refactor(plugin): extract vulkan_kernels.rs (graphics + ray-tracing combined) (#1079)
Combines the VulkanGraphicsKernelMethodsVTable and
VulkanRayTracingKernelMethodsVTable callback bodies into a single
sibling submodule. Splitting them would require surgically picking
apart heavily interleaved Linux + non-Linux platform-stub variants
of each method (the source layout intermixes both kernel families
across the same 1.8k-line section), and the runtime shape is the
same: reconstruct kernel borrow → run inner method →
i32 + err_buf return shape. Combining them keeps the boundary
clean and the file size in the same range as the other extracted
submodules. Note the deviation from the issue's per-vtable sketch.

mod.rs re-exports HOST_VULKAN_GRAPHICS_KERNEL_METHODS_VTABLE,
HOST_VULKAN_RAY_TRACING_KERNEL_METHODS_VTABLE,
host_vulkan_graphics_kernel_methods_vtable, and
host_vulkan_ray_tracing_kernel_methods_vtable. Both per-vtable
test modules (graphics_kernel_methods_vtable_null_tests +
ray_tracing_kernel_methods_vtable_null_tests) move with the body.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
